Initial commit
Very basic walking skeleton demonstrates two colors with two pawn each. Home is the 10th space. Every card is a 1.

Future commits will add more complexity and the ability to simulate a large number of games in order to identify the benefit of position in a non-symmetrical game and different decisions